Created custom reporting
Added ability to create custom reports. This will replace the pages specifically created for reports. We may need to resurrect some of them if we choose to create alerts.

Additional reports can be added, edited and the sort order changed.

REQUIRES ADDITIONAL TABLE - Haven't added code to create it yet...