Allow for fixing of possible updates on report
New links on the SoftwareMatrix Update Report allow for the fixing of possible updates listed that need verification. If there is no update available, the user can click a link and confirm that it was reviewed and the database will be updated with the new location of the version on the web page.